requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
google-genai
vertexai
google-generativeai
rapidfuzz